pyproject.toml
[tool.poetry]
name = "msc"
version = "0.1.0"
description = ""
authors = ["Jack Collis <jackcollis96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
fastapi = "^0.103.1"
python = "^3.9"
sqlalchemy = "^2.0.20"
sqlalchemy-utils = "^0.41.1"
pydantic = "1.10.2"
alembic = "^1.11.1"
uvicorn = "^0.23.1"
psycopg2 = "^2.9.7"
requests = "^2.31.0"
boto3 = "^1.28.46"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
black = "^23.7.0"
isort = "^5.12.0"
autopep8 = "^2.0.2"
httpx = "^0.24.1"
freezegun = "^1.2.2"

[tool.poetry.scripts]
start_docker = "scripts.docker_compose_up:main"
build_api = "scripts.build_api:main"
deploy_api = "scripts.deploy_api:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
